+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/MultibindsChecker.kt
@@ -6,6 +6,7 @@ import dev.zacsweers.metro.compiler.fir.MetroDiagnostics
 import dev.zacsweers.metro.compiler.fir.MetroFirAnnotation
 import dev.zacsweers.metro.compiler.fir.annotationsIn
 import dev.zacsweers.metro.compiler.fir.classIds
+import dev.zacsweers.metro.compiler.fir.diagnosticString
 import dev.zacsweers.metro.compiler.fir.hasImplicitClassKey
 import dev.zacsweers.metro.compiler.fir.isOrImplements
 import dev.zacsweers.metro.compiler.fir.mapKeyClassValueExpression
@@ -169,7 +170,7 @@ internal object MultibindsChecker : FirCallableDeclarationChecker(MppCheckerKind
                         reporter.reportOn(
                           declaration.returnTypeRef.source ?: source,
                           MetroDiagnostics.MULTIBINDS_ERROR,
-                          "Multibinding map key '${keyClass.classId.asFqNameString()}' is not annotated with @MapKey(unwrapValue = false).",
+                          "Multibinding map key '${keyClass.classId.diagnosticString}' is not annotated with @MapKey(unwrapValue = false).",
                         )
                       } else if (
                         mapKey.getBooleanArgument(Symbols.Names.unwrapValue, session) != false
@@ -177,7 +178,7 @@ internal object MultibindsChecker : FirCallableDeclarationChecker(MppCheckerKind
                         reporter.reportOn(
                           declaration.returnTypeRef.source ?: source,
                           MetroDiagnostics.MULTIBINDS_ERROR,
-                          "Multibinding map key '${keyClass.classId.asFqNameString()}' is annotated with @MapKey but does not set MapKey.unwrapValue to false. This is necessary to use wrapped values as multibinding map keys.",
+                          "Multibinding map key '${keyClass.classId.diagnosticString}' is annotated with @MapKey but does not set MapKey.unwrapValue to false. This is necessary to use wrapped values as multibinding map keys.",
                         )
                       }
                     } else {
@@ -336,7 +337,7 @@ internal fun checkImplicitClassKeyUsage(
     reporter.reportOn(
       argSource,
       MetroDiagnostics.MAP_KEY_REDUNDANT_IMPLICIT_CLASS_KEY,
-      "Explicit class key value '${valueClassId.asFqNameString()}::class' is the same as the implicit class key and can be omitted.",
+      "Explicit class key value '${valueClassId.diagnosticString}::class' is the same as the implicit class key and can be omitted.",
     )
   }
 }