+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/injectionSiteChecks.kt
@@ -119,7 +119,7 @@ internal fun validateInjectionSiteType(
       reporter.reportOn(
         typeRef.source ?: source,
         MetroDiagnostics.SUSPICIOUS_OBJECT_INJECTION_WARNING,
-        "Suspicious injection of an unqualified object type '${clazz.classId.asFqNameString()}'. This is probably unnecessary or unintentional.",
+        "Suspicious injection of an unqualified object type '${clazz.classId.diagnosticString}'. This is probably unnecessary or unintentional.",
       )
     } else {
       val isAssistedInject =
@@ -138,16 +138,16 @@ internal fun validateInjectionSiteType(
               ?.symbol
 
         val message = buildString {
-          val fqName = clazz.classId.asFqNameString()
+          val fqName = clazz.classId.diagnosticString
           append(
-            "[Metro/InvalidBinding] '$fqName' uses assisted injection and cannot be injected directly into 'test.ExampleGraph.exampleClass'. You must inject a corresponding @AssistedFactory type or provide a qualified instance on the graph instead."
+            "'$fqName' uses assisted injection and cannot be injected directly here. You must inject a corresponding @AssistedFactory type or provide a qualified instance on the graph instead."
           )
           if (nestedFactory != null) {
             appendLine()
             appendLine()
             appendLine("(Hint)")
             appendLine(
-              "It looks like the @AssistedFactory for '$fqName' may be '${nestedFactory.classId.asFqNameString()}'."
+              "It looks like the @AssistedFactory for '$fqName' may be '${nestedFactory.classId.diagnosticString}'."
             )
           }
         }
@@ -157,7 +157,7 @@ internal fun validateInjectionSiteType(
           message,
         )
       } else if (clazz.usesContributionProviderPath(session)) {
-        val fqName = clazz.classId.asFqNameString()
+        val fqName = clazz.classId.diagnosticString
         reporter.reportOn(
           typeRef.source ?: source,
           MetroDiagnostics.NON_EXPOSED_IMPL_TYPE,
@@ -241,7 +241,7 @@ private fun checkLazyAssistedFactory(
       typeRef.source ?: source,
       MetroDiagnostics.ASSISTED_FACTORIES_CANNOT_BE_LAZY,
       canonicalClass.name.asString(),
-      canonicalClass.classId.asFqNameString(),
+      canonicalClass.classId.diagnosticString,
     )
   }
 }